move #private methods on agent.js to static functions
